Benchmark tools: Remove the governor configuration
The governor configuration is not appropriate, since what needs to be done is to lock the CPU clock.

However, it seems that there isn't a stable software way to do that; on a Ryzen 3950x, the strategies:

- setting `userspace` governor
- using `cpufreq-set`
- using `cpupower`
- setting `/sys/**/cpufreq/scaling_max_freq`

didn't work (verified via `grep MHz /proc/cpuinfo`), so the previous strategy has been replaced with a warning in the help.